ResourceAccessorELink: matched getMtxSetType

+++ include/xlink2/xlink2ResourceAccessorELink.h
@@ -33,7 +33,7 @@ public:
     bool isForceLoopAsset(const ResAssetCallTable& asset_ctb) const;
 
     s32 getForceCalc(const ResAssetCallTable& asset_ctb) const;
-    u64 getMtxSetType(const ResAssetCallTable& asset_ctb) const;
+    s32 getMtxSetType(const ResAssetCallTable& asset_ctb) const;
     f64 getRotateSourceType(const ResAssetCallTable& asset_ctb) const;
 
     bool isFollow(const ResAssetCallTable& asset_ctb) const;
